requirements/base.txt
Django>=2.0.1
django-modeltranslation==0.13b1
psycopg2==2.7.4
python-dateutil==2.6.1
pylint>=2.1.1
pylint-django>=2.0